requirements.txt
# Core dependencies
openai>=1.0.0
azure-ai-documentintelligence>=1.0.0
azure-core>=1.29.0
python-dotenv>=1.0.0
chromadb>=0.4.0
pandas>=1.5.0
tiktoken>=0.5.1